Implementação de mais testes na camada de controller

+++ src/main/java/br/com/empresa/controller/FuncionarioController.java
@@ -2,6 +2,8 @@ package br.com.empresa.controller;
 
 import br.com.empresa.model.FuncionarioModel;
 import br.com.empresa.service.FuncionarioService;
+import br.com.empresa.util.LoggerUtil;
+import br.com.empresa.util.ValidacoesUtil;
 
 import java.util.List;
 
@@ -13,18 +15,34 @@ public class FuncionarioController {
     }
 
     public int criarFuncionario(FuncionarioModel funcionario){
+        if(funcionario == null || ValidacoesUtil.validaFuncionario(funcionario)){
+            LoggerUtil.logInfo("[FUNCIONARIO CONTROLLER] FUNCIONARIO INVALIDO");
+            return -1;
+        }
         return funcionarioService.inserir(funcionario);
     }
 
     public boolean atualizarFuncionario(FuncionarioModel funcionario){
+        if(funcionario == null || ValidacoesUtil.validaFuncionario(funcionario)){
+            LoggerUtil.logInfo("[FUNCIONARIO CONTROLLER] FUNCIONARIO INVALIDO");
+            return false;
+        }
         return funcionarioService.atualizar(funcionario);
     }
 
     public boolean removerFuncionario(int id){
+        if (!ValidacoesUtil.validaID(id)) {
+            LoggerUtil.logWarning("ID NÃO VALIDO. ID: " + id);
+            return false;
+        }
         return funcionarioService.remover(id);
     }
 
     public FuncionarioModel buscarFuncionarioPorId(int id){
+        if (!ValidacoesUtil.validaID(id)) {
+            LoggerUtil.logWarning("ID NÃO VALIDO. ID: " + id);
+            return null;
+        }
         return funcionarioService.buscaPorId(id);
     }
